Inputs refactoring

+++ src/views/managers/template.svelte
@@ -9,7 +9,6 @@
     AddPropModal,
     NoEntities,
     NameDescModal,
-    RemoveManagerModal,
   } from '../../components';
   import {
     EntityManagerService,
@@ -81,14 +80,24 @@
     );
   }
   async function remove() {
-    await GeneralService.errorWrapper(
-      async () => {
-        await EntityManagerService.delete('template', template._id);
-      },
-      async () => {
-        NotificationService.success('Template was successfully deleted.');
-      }
-    );
+    if (
+      await ConfirmService.confirm(
+        `Delete "${template.label}" Template`,
+        `Are you sure you want to delete <strong>${template.label}</strong>
+          template? This action is irreversible and all entries in
+          this template will also be deleted.`,
+        template.name
+      )
+    ) {
+      await GeneralService.errorWrapper(
+        async () => {
+          await EntityManagerService.delete('template', template._id);
+        },
+        async () => {
+          NotificationService.success('Template was successfully deleted.');
+        }
+      );
+    }
   }
   async function addProp(prop: Prop) {
     await GeneralService.errorWrapper(
@@ -214,7 +223,7 @@
             updateProp(event.detail);
           }}
           on:deleteEntity={() => {
-            StoreService.update('RemoveManagerModal', true);
+            remove();
           }}
           on:deleteProp={(event) => {
             removeProp(event.detail);
@@ -254,13 +263,4 @@
         create(event.detail.name, event.detail.desc);
       }
     }} />
-  {#if template}
-    <RemoveManagerModal
-      title={`Delete "${template.label}" Template`}
-      warningMessage={`Are you sure you want to delete "${template.label}" template?
-  This action is irreversible and all entries in this template will also be deleted.`}
-      inputLabel="Confirm template name"
-      managerName={template.label}
-      on:done={remove} />
-  {/if}
 </Layout>
